+++ package.json
@@ -21,7 +21,7 @@
     "postcss": "^8.4.18",
     "prettier-plugin-svelte": "^3.2.3",
     "sass": "^1.55.0",
-    "svelte": "5.0.0-next.120",
+    "svelte": "5.0.0-next.123",
     "svelte-check": "^3.7.0",
     "typescript": "^5.4.5",
     "vite": "5.2.10",
